chore: move files from utils to src/utils; move files within utils/validation to src/middleware for better file structuring; move passport-config.js to src/config
